Move dutytracer into exporter2
