+++ .gitattributes
@@ -1,1 +1,1 @@
-operator/dutytracer/benchdata/slot_3707881_3707882.ssz filter=lfs diff=lfs merge=lfs -text
+exporter2/dutytracer/benchdata/slot_3707881_3707882.ssz filter=lfs diff=lfs merge=lfs -text
