+++ exporter2/dutytracer/benchdata/slot_3707881_3707882.ssz
