+++ exporter2/dutytracer/collector.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"bytes"
