+++ exporter2/dutytracer/collector_bench_test.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"context"
