+++ exporter2/dutytracer/collector_quorum_test.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"context"
