+++ exporter2/dutytracer/collector_schedule_test.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"testing"
