+++ exporter2/dutytracer/eviction.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"encoding/hex"
