+++ exporter2/dutytracer/store.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"errors"
