+++ exporter2/dutytracer/store_metrics.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"context"
