+++ exporter2/dutytracer/store_test.go
@@ -1,4 +1,4 @@
-package validator
+package dutytracer
 
 import (
 	"encoding/binary"
